refactor(run): extract shared run formatting helpers into src/runFormat.ts
Closes #87
